+++ package.json
@@ -16,7 +16,8 @@
     "react": "^18",
     "react-dom": "^18",
     "react-responsive": "^10.0.0",
-    "react-tailwindcss-datepicker": "^1.6.6"
+    "react-tailwindcss-datepicker": "^1.6.6",
+    "responsive": "^2.0.0"
   },
   "devDependencies": {
     "@types/node": "^20",
