+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Header from "./header";
-import Decor from "./decor";
+import Decor from "../components/decor";
 
 import localFont from "next/font/local";
