+++ src/components/background.tsx
