+++ src/components/decor.tsx
